Minimal working reader
Working demo of the reader. There are no settings. It just shows first
Chapter of Genji Monogatari and opens a dictionary popup for selected
words. 90% of the code is dumped into a single html file. Styles need a
lot of work. Should work on Android 4.0.4 and up. Can be tested on
Chrome with device emulation. Should even work on FireFox, but touch
events won't trigger.