Possibly solved select triggering randomly on T68
While reading on T68, sometimes dictionary does not show up. Then on
next screen touch a piece of text gets selected as with a long tap. This
might be caused by touchend event not triggering. The solution I used
was forcing touchend event on cancelled touch in the reader container.